MIM-50 拆分订阅回归测试并同步 main
